[CORE-6083]
