+++ projects/v3/src/app/pages/devtool/devtool.page.ts
@@ -4,6 +4,7 @@ import { ExperienceService } from '@v3/app/services/experience.service';
 import { FastFeedbackService } from '@v3/app/services/fast-feedback.service';
 import { NotificationsService } from '@v3/app/services/notifications.service';
 import { BrowserStorageService } from '@v3/app/services/storage.service';
+import { SharedService } from '@v3/app/services/shared.service';
 
 @Component({
   selector: 'app-devtool',
@@ -20,6 +21,7 @@ export class DevtoolPage implements OnInit {
     private fastFeedbackService: FastFeedbackService,
     private notificationsService: NotificationsService,
     private experienceService: ExperienceService,
+    private sharedService: SharedService,
   ) { }
 
   ngOnInit() {
@@ -30,7 +32,7 @@ export class DevtoolPage implements OnInit {
   }
 
   refresh() {
-    this.experienceService.getNewJwt().subscribe();
+    this.sharedService.getNewJwt().subscribe();
   }
 
   login() {
